requirements.txt
streamlit>=1.31,<2.0
openai>=1.12,<2.0
google-generativeai>=0.3.0
pandas>=2.2,<3.0
python-dotenv>=1.0,<2.0
tenacity>=8.2,<9.0
requests>=2.31,<3.0
selenium>=4.17,<5.0
beautifulsoup4>=4.12,<5.0
webdriver-manager>=4.0,<5.0
yt-dlp>=2024.3.10
